requirements.txt
pytest
pytest-dotenv
pytest-mock
rich
